refactor(listener): extract branch and secret commands
